+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1]
 
-* TODO: Describe initial release.
+* added intial release
+
+## [0.0.2]
+
+* added numeric values parser
+
+## [0.0.3]
+
+* added language local cache support + bug fixes
+
+## [0.0.4]
+
+* fixed dependency issue